+++ spec/mongoid/relations/embedded/many_spec.rb
@@ -3254,4 +3254,38 @@ describe Mongoid::Relations::Embedded::Many do
       criteria.should eq([ address ])
     end
   end
+
+  context "when updating multiple levels in one update" do
+
+    let!(:person) do
+      Person.create(
+        addresses: [
+          { locations: [{ name: "home" }]}
+        ]
+      )
+    end
+
+    context "when updating with hashes" do
+
+      let(:from_db) do
+        Person.find(person.id)
+      end
+
+      before do
+        from_db.update_attributes(
+          addresses: [
+            { locations: [{ name: "work" }]}
+          ]
+        )
+      end
+
+      let(:updated) do
+        person.reload.addresses.first.locations.first
+      end
+
+      it "updates the nested document" do
+        updated.name.should eq("work")
+      end
+    end
+  end
 end